+++ .env
@@ -1,1 +1,1 @@
-REACT_APP_CRUD_CRUD=https://crudcrud.com/api/fd127fa38d26470cb77dbbf8fdc1a763
+REACT_APP_CRUD_CRUD=https://crudcrud.com/api/b427d9e567264f11a25bec1104f2f517
