+++ src/App.tsx
@@ -2,11 +2,11 @@ import React from "react";
 import "./index.css";
 import { Provider } from "react-redux";
 import store from "./state/Configurestore";
-import StudentTableContainer from "./containers/StudentTableContainer";
 import "./App.css";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import StudentFormContainer from "./containers/StudentFormContainer";
 import RoutingHandler from "./helpers/history";
+import StudentDashboard from "./components/studentDashboard";
 
 function App() {
   return (
@@ -15,7 +15,7 @@ function App() {
         <BrowserRouter>
           <RoutingHandler />
           <Routes>
-            <Route path="/" element={<StudentTableContainer />} />
+            <Route path="/" element={<StudentDashboard />} />
             <Route path="student" element={<StudentFormContainer />} />
             <Route
               path="student/:studentId"
