+++ src/state/ducks/student/types.ts
@@ -2,7 +2,6 @@ import { MetaAction, TypeConstant } from "typesafe-actions";
 
 export interface IStudentState {
   list: IStudentRaw[];
-  summary: Record<string, string>;
 }
 export interface IStudentStateModified {
   list: IStudentRaw[];
